backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
faker>=19.0.0
bcrypt<4.0.0
